Changes to have Decimal on Altitude Measure and Tracks recording in order to - Get benefit of higher resolution measure with Barometer - Make post treatment applications less confused by not having 1m jump between 2 consecutive record
==> Include Barometer based Decimal Altitude

+++ app/src/main/java/ch/bailu/aat/services/sensor/internal/BarometerSensor.java
@@ -111,10 +111,10 @@ public class BarometerSensor extends InternalSensorSDK23
     public static class Information extends GpxInformation {
 
         private final GpxAttributes attributes;
-        private final short altitude;
+        private final double altitude;
         private final long time = System.currentTimeMillis();
 
-        public Information(short a, float p) {
+        public Information(double a, float p) {
             attributes = new Attributes(p);
             altitude = a;
         }
@@ -128,7 +128,7 @@ public class BarometerSensor extends InternalSensorSDK23
 
         @Override
         public double getAltitude() {
-            return altitude;
+            return  altitude;
         }
